requirements.txt
amqp==5.2.0
asgiref==3.7.2
attrs==23.1.0
billiard==4.2.0
celery==5.3.5
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.7
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-filter==23.3
django-redis==3.1.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
drf-yasg==1.21.7
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
kombu==5.3.4
packaging==23.2
prompt-toolkit==3.0.41
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.31.0
rest-framework-simplejwt==0.0.2
rpds-py==0.13.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.12
